runtime: Split into Runtime interface and StreamRuntime implementation
