+++ src/main/java/com/abusalimov/mrcalc/runtime/Runtime.java
@@ -1,10 +1,6 @@
 package com.abusalimov.mrcalc.runtime;
 
 import java.util.function.*;
-import java.util.stream.DoubleStream;
-import java.util.stream.LongStream;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 /**
  * The runtime provides the necessary support functions (mainly for map() / reduce()) used for evaluating an expression.
@@ -12,25 +8,7 @@ import java.util.stream.StreamSupport;
  *
  * @author Eldar Abusalimov
  */
-public class Runtime {
-    private final boolean parallel;
-
-    /**
-     * Creates a new instance providing functions executing in parallel, where possible.
-     */
-    public Runtime() {
-        this(true);
-    }
-
-    /**
-     * Creates a new instance operating in parallel, as indicated by the argument.
-     *
-     * @param parallel whether to use parallel operations, where possible, or not
-     */
-    public Runtime(boolean parallel) {
-        this.parallel = parallel;
-    }
-
+public interface Runtime {
     /**
      * Creates a new {@link LongSequence} filled by integers between the specified boundaries.
      *
@@ -38,71 +16,158 @@ public class Runtime {
      * @param endExclusive   the end boundary of the range (exclusive)
      * @return the new {@link LongSequence} instance
      */
-    public LongSequence createLongRange(long startInclusive, long endExclusive) {
-        return new LongRange(startInclusive, endExclusive);
-    }
+    LongSequence createLongRange(long startInclusive, long endExclusive);
 
-    public <E> E reduce(Sequence<E> sequence, E identity, BinaryOperator<E> operator) {
-        return objectStream(sequence).reduce(identity, operator);
-    }
+    /**
+     * Performs a reduction on the elements of the given sequence, using the provided identity value and an associative
+     * accumulation function, and returns the reduced value.
+     * <p>
+     * All values involved into the reduction, as well as the return type, are treated as generic objects, and
+     * primitives must be boxed appropriately. In case of using a sequence of primitives, to achieve better performance,
+     * a specialized method should be considered instead.
+     *
+     * @param sequence the sequence of objects to reduce
+     * @param identity the neutral element, which must be an identity for the operator
+     * @param operator the accumulator function: {@code E, E -> E}
+     * @param <E>      a type of the values, involved into the reduction as well as the return type
+     * @return the result of the reduction
+     * @see #reduceLong(LongSequence, long, LongBinaryOperator) for reduction of the primitive longs
+     * @see #reduceDouble(DoubleSequence, double, DoubleBinaryOperator) for reduction of the primitive doubles
+     */
+    <E> E reduce(Sequence<E> sequence, E identity, BinaryOperator<E> operator);
 
-    public long reduceLong(LongSequence sequence, long identity, LongBinaryOperator operator) {
-        return longStream(sequence).reduce(identity, operator);
-    }
+    /**
+     * Performs a reduction on the elements of the given sequence of primitive longs, using the provided long identity
+     * value and an associative accumulation function, and returns the reduced value.
+     *
+     * @param sequence the sequence of primitive longs to reduce
+     * @param identity the neutral element, which must be an identity for the operator
+     * @param operator the accumulator function: {@code long, long -> long}
+     * @return the result of the reduction
+     */
+    long reduceLong(LongSequence sequence, long identity, LongBinaryOperator operator);
 
-    public double reduceDouble(DoubleSequence sequence, double identity, DoubleBinaryOperator operator) {
-        return doubleStream(sequence).reduce(identity, operator);
-    }
+    /**
+     * Performs a reduction on the elements of the given sequence of primitive doubles, using the provided double
+     * identity value and an associative accumulation function, and returns the reduced value.
+     *
+     * @param sequence the sequence of primitive doubles to reduce
+     * @param identity the neutral element, which must be an identity for the operator
+     * @param operator the accumulator function: {@code double, double -> double}
+     * @return the result of the reduction
+     */
+    double reduceDouble(DoubleSequence sequence, double identity, DoubleBinaryOperator operator);
 
-    @SuppressWarnings("unchecked")
-    public <E, R> Sequence<R> mapToObject(Sequence<E> sequence, Function<? super E, ? extends R> mapper) {
-        return new ObjectSequence(objectStream(sequence).map(mapper).toArray());
-    }
+    /**
+     * Returns a sequence consisting of the results of applying the given function to the elements of the specified
+     * sequence.
+     * <p>
+     * All values are treated as generic objects, and primitives must be boxed appropriately. In case of using a
+     * sequence of primitives, to achieve better performance, a specialized method should be considered instead.
+     *
+     * @param sequence the sequence of objects to map
+     * @param mapper   the function to apply for each element: {@code E -> R}
+     * @param <E>      the element type of the source sequence
+     * @param <R>      the element type of the destination sequence
+     * @return the new sequence of objects
+     * @see #mapLongToObject(LongSequence, LongFunction) for mapping sequences of primitive longs
+     * @see #mapDoubleToObject(DoubleSequence, DoubleFunction) for mapping sequences of primitive doubles
+     */
+    <E, R> Sequence<R> mapToObject(Sequence<E> sequence, Function<? super E, ? extends R> mapper);
 
-    @SuppressWarnings("unchecked")
-    public <R> Sequence<R> mapLongToObject(LongSequence sequence, LongFunction<? extends R> mapper) {
-        return new ObjectSequence(longStream(sequence).mapToObj(mapper).toArray());
-    }
+    /**
+     * Returns a sequence consisting of the results of applying the given function to the elements of the specified
+     * sequence of primitive longs.
+     *
+     * @param sequence the sequence of primitive longs to map
+     * @param mapper   the function to apply for each element: {@code long -> R}
+     * @param <R>      the element type of the destination sequence
+     * @return the sequence of objects
+     */
+    <R> Sequence<R> mapLongToObject(LongSequence sequence, LongFunction<? extends R> mapper);
 
-    @SuppressWarnings("unchecked")
-    public <R> Sequence<R> mapDoubleToObject(DoubleSequence sequence, DoubleFunction<? extends R> mapper) {
-        return new ObjectSequence(doubleStream(sequence).mapToObj(mapper).toArray());
-    }
+    /**
+     * Returns a sequence consisting of the results of applying the given function to the elements of the specified
+     * sequence of primitive doubles.
+     *
+     * @param sequence the sequence of primitive doubles to map
+     * @param mapper   the function to apply for each element: {@code double -> R}
+     * @param <R>      the element type of the destination sequence
+     * @return the sequence of objects
+     */
+    <R> Sequence<R> mapDoubleToObject(DoubleSequence sequence, DoubleFunction<? extends R> mapper);
 
-    public <E> LongSequence mapToLong(Sequence<E> sequence, ToLongFunction<? super E> mapper) {
-        return new LongSequence(objectStream(sequence).mapToLong(mapper).toArray());
-    }
+    /**
+     * Returns a sequence of primitive longs consisting of the results of applying the given function to the elements of
+     * the specified sequence.
+     * <p>
+     * The elements of the source sequence are treated as generic objects, and primitives must be boxed appropriately.
+     * In case of using a sequence of primitives, to achieve better performance, a specialized method should be
+     * considered instead.
+     *
+     * @param sequence the sequence of objects to map
+     * @param mapper   the function to apply for each element: {@code E -> long}
+     * @param <E>      the element type of the source sequence
+     * @return the sequence of primitive longs
+     * @see #mapLongToLong(LongSequence, LongUnaryOperator) for mapping sequences of primitive longs
+     * @see #mapDoubleToLong(DoubleSequence, DoubleToLongFunction) for mapping sequences of primitive doubles
+     */
+    <E> LongSequence mapToLong(Sequence<E> sequence, ToLongFunction<? super E> mapper);
 
-    public LongSequence mapLongToLong(LongSequence sequence, LongUnaryOperator mapper) {
-        return new LongSequence(longStream(sequence).map(mapper).toArray());
-    }
+    /**
+     * Returns a sequence of primitive longs consisting of the results of applying the given function to the elements of
+     * the specified sequence of primitive longs.
+     *
+     * @param sequence the sequence of primitive longs to map
+     * @param mapper   the function to apply for each element: {@code long -> long}
+     * @return the sequence of primitive longs
+     */
+    LongSequence mapLongToLong(LongSequence sequence, LongUnaryOperator mapper);
 
-    public LongSequence mapDoubleToLong(DoubleSequence sequence, DoubleToLongFunction mapper) {
-        return new LongSequence(doubleStream(sequence).mapToLong(mapper).toArray());
-    }
+    /**
+     * Returns a sequence of primitive longs consisting of the results of applying the given function to the elements of
+     * the specified sequence of primitive doubles.
+     *
+     * @param sequence the sequence of primitive doubles to map
+     * @param mapper   the function to apply for each element: {@code double -> long}
+     * @return the sequence of primitive longs
+     */
+    LongSequence mapDoubleToLong(DoubleSequence sequence, DoubleToLongFunction mapper);
 
-    public <E> DoubleSequence mapToDouble(Sequence<E> sequence, ToDoubleFunction<? super E> mapper) {
-        return new DoubleSequence(objectStream(sequence).mapToDouble(mapper).toArray());
-    }
+    /**
+     * Returns a sequence of primitive doubles consisting of the results of applying the given function to the elements
+     * of the specified sequence.
+     * <p>
+     * The elements of the source sequence are treated as generic objects, and primitives must be boxed appropriately.
+     * In case of using a sequence of primitives, to achieve better performance, a specialized method should be
+     * considered instead.
+     *
+     * @param sequence the sequence of objects to map
+     * @param mapper   the function to apply for each element: {@code E -> double}
+     * @param <E>      the element type of the source sequence
+     * @return the sequence of primitive doubles
+     * @see #mapLongToDouble(LongSequence, LongToDoubleFunction) for mapping sequences of primitive longs
+     * @see #mapDoubleToDouble(DoubleSequence, DoubleUnaryOperator) for mapping sequences of primitive doubles
+     */
+    <E> DoubleSequence mapToDouble(Sequence<E> sequence, ToDoubleFunction<? super E> mapper);
 
-    public DoubleSequence mapLongToDouble(LongSequence sequence, LongToDoubleFunction mapper) {
-        return new DoubleSequence(longStream(sequence).mapToDouble(mapper).toArray());
-    }
+    /**
+     * Returns a sequence of primitive doubles consisting of the results of applying the given function to the elements
+     * of the specified sequence of primitive longs.
+     *
+     * @param sequence the sequence of primitive longs to map
+     * @param mapper   the function to apply for each element: {@code long -> double}
+     * @return the sequence of primitive doubles
+     */
+    DoubleSequence mapLongToDouble(LongSequence sequence, LongToDoubleFunction mapper);
 
-    public DoubleSequence mapDoubleToDouble(DoubleSequence sequence, DoubleUnaryOperator mapper) {
-        return new DoubleSequence(doubleStream(sequence).map(mapper).toArray());
-    }
-
-    protected <E> Stream<E> objectStream(Sequence<E> sequence) {
-        return StreamSupport.stream(sequence.spliterator(), parallel);
-    }
-
-    protected LongStream longStream(LongSequence sequence) {
-        return StreamSupport.longStream(sequence.spliterator(), parallel);
-    }
-
-    protected DoubleStream doubleStream(DoubleSequence sequence) {
-        return StreamSupport.doubleStream(sequence.spliterator(), parallel);
-    }
-
+    /**
+     * Returns a sequence of primitive doubles consisting of the results of applying the given function to the elements
+     * of the specified sequence of primitive doubles.
+     *
+     * @param sequence the sequence of primitive doubles to map
+     * @param mapper   the function to apply for each element: {@code double -> double}
+     * @return the sequence of primitive doubles
+     */
+    DoubleSequence mapDoubleToDouble(DoubleSequence sequence, DoubleUnaryOperator mapper);
 }
