+++ src/test/groovy/com/abusalimov/mrcalc/InterpreterTest.groovy
@@ -6,6 +6,7 @@ import com.abusalimov.mrcalc.compile.Compiler
 import com.abusalimov.mrcalc.parse.Parser
 import com.abusalimov.mrcalc.parse.impl.antlr.ANTLRParserImpl
 import com.abusalimov.mrcalc.runtime.Runtime
+import com.abusalimov.mrcalc.runtime.StreamRuntime
 import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
@@ -27,7 +28,7 @@ class InterpreterTest {
         parser = new ANTLRParserImpl()
         backend = new FuncBackendImpl()
         compiler = new Compiler(backend)
-        runtime = new Runtime()
+        runtime = new StreamRuntime()
         interpreter = new Interpreter(runtime)
     }
